+++ CHANGELOG.md
@@ -12,9 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Added CMIP6 data source
 - Added CBottle Tropical Cyclone guidance diagnostic
-- Added ModelOutputDatasetSource to use written model output to start a new model run
+- Added CBottle Video prognostic model
 - Exposed backend arguments of netcdf/zarr to datasource_to_file signature
 - Added vertical wind speed support in GFS
+- Added ModelOutputDatasetSource to use written model output to start a new model run
 
 ### Changed